requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
dotenv==0.9.9
idna==3.11
pillow==11.3.0
python-dotenv==1.2.1
requests==2.32.5
urllib3==1.26.15
